recommendation/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pinecone==6.0.2
python-dotenv==1.0.1
sentence-transformers==2.6.0
pydantic==2.11.4
pydantic-settings==2.9.1 